backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
httpx>=0.25.0
python-dotenv>=1.0.0
anthropic>=0.40.0
pydantic>=2.5.0
PyJWT[crypto]>=2.8.0
python-docx>=1.1.0
tqdm>=4.66.0
tabulate>=0.9.0
